+++ acoustic/python/bluerock/__main__.py
@@ -1,7 +1,7 @@
 # Copyright (C) 2025 BlueRock Security, Inc.
 # All rights reserved.
 
-from . import common
+from bluepython import common
 
 if __name__ == "__main__":
     common.main()